Construct minimal set of parameters to least_squares for doses.
This reduces out any parameters that have a 'fixed' bound so that they
are not present in the X vector that is optimized. This is a big
improvement over using infinitesimal bounds to represent fixed
parameters, and the optimization runs a lot better and a lot faster this
way.

It also renames the function to fit.createX0AndBounds.